move ElevationProfileChip to chips module
