+++ src/modules/chips/components/assistChips/ElevationProfileChip.js
@@ -3,7 +3,7 @@
  */
 import { $injector } from '../../../../injection';
 import { openProfile } from '../../../../store/elevationProfile/elevationProfile.action';
-import { AbstractAssistChip } from '../../../chips/components/assistChips/AbstractAssistChip';
+import { AbstractAssistChip } from './AbstractAssistChip';
 import profileSvg from './assets/profile.svg';
 
 const Update_Profile_Coordinates = 'update_profile_coordinates';
