+++ src/modules/chips/components/assistChips/index.js
@@ -1,0 +1,4 @@
+import { ElevationProfileChip } from './ElevationProfileChip';
+if (!window.customElements.get(ElevationProfileChip.tag)) {
+	window.customElements.define(ElevationProfileChip.tag, ElevationProfileChip);
+}
